+++ src/main/java/store/myproject/onlineshop/domain/membership/dto/MemberShipUpdateRequest.java
@@ -21,7 +21,7 @@ public class MemberShipUpdateRequest {
     private BigDecimal baseline;
 
     @NotNull
-    private Float discountRate;
+    private BigDecimal discountRate;
 
     public MemberShip toEntity() {
         return MemberShip.builder()
